requirements.txt
streamlit>=1.35.0
requests>=2.31.0
python-docx>=1.1.0
pyyaml>=6.0
openai>=1.30.0
